Path:  Now testing with header by default.
Modified the tests to leave out the "--no-header" option.
Renamed the "metric" tests to "default" tests.
Added a "no_header" test.
Fixed the original centroid postprocessor to use the basename
for the cam file instead of the whole path, which will vary
from system to system.
Modified the testing routine to ignore the "Output Time:" lines.